requirements.txt
albumentations==1.1.0
numpy==1.20.3
opencv_python_headless==4.5.5.62
pandas==1.3.4
Pillow==9.2.0
pytorch_lightning==1.5.10
PyYAML==6.0
requests==2.26.0
scikit_learn==1.1.1
scipy==1.7.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3